finished profile page mostly, need to add feed component
